Restore null scope check to stop EventBridge injection (#7919)
## Summary of changes

When the EventBridge instrumentation is disabled we should not inject
headers / context into the object. This was the case initially in
https://github.com/DataDog/dd-trace-dotnet/pull/6096, but appears to
have been accidentally removed in a refactoring in
https://github.com/DataDog/dd-trace-dotnet/pull/6157.

## Reason for change

Appears that we do not entirely support certain message types here.

## Implementation details

Essentially restored the `if (scope?.Span.Context is { } context)` check
before injecting.

## Test coverage

Our test coverage _passes_ with / without this as we correctly don't
create a span when the instrumentation is disabled but we don't really
have the means to assert that we do not modify the underlying objects in
some way. This is potentially something we should look into
implementation to catch future issues like this potentially.

## Other details
<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->