Add some documentation on updating Windows GitLab image for CI (#7782)
## Summary of changes

Adds some additional documentation around updating the Windows GitLab
image, we needed to do this recently and the documentation for it was a
bit sparse, and we've had to do it also for the .NET updates recently.

## Reason for change

Try to add some documentation on how to do it.

## Implementation details

I based this off of our previous example PRs, some slack messages, and
existing documentation and asked Claude to do the rest.

## Test coverage

🙈 

## Other details
<!-- Fixes #{issue} -->

https://datadoghq.atlassian.net/browse/LANGPLAT-790

I figure we can try to follow this in the .NET 10 upgrade next week to
see if the steps are accurate as I've never actually done them 😛

<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->

---------

Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>