Don't use `System.Runtime.CompilerServices.Unsafe` in .NET Core 2.1 (#7955)
## Summary of changes

Uses the vendored `System.Runtime.CompilerServices.Unsafe` instead of
the BCL version

## Reason for change

The BCL version is not available in .NET Core 2.1. If it's not made
available by some other means (e.g. by aspnetcore), then spans won't
close correctly.

However, we were previously swallowing exceptions thrown when a `Scope`
is disposed, so we had no visibility on this. I can't see a good reason
to do that, so changed to log an exception instead (arguably, we
could/should remove the try-catch entirely and catch in the integration,
but this seems safer in the short term).

## Implementation details

- Use the vendored unsafe in < .NET Core 3.1
- Add a log in `Scope.Dispose()`

## Test coverage

Confirmed the assembly ref has gone, and did manual testing with .NET
Core 2.1, but as we don't support or test that directly, there's no
explicit tests for it in CI

## Other details

We should probably review our other swallowing instances, to make sure
we're not missing anything else important