bump libdatadog version to v25 (#7971)
## Summary of changes

just a bump, I need this new version to pass process tags with profiles.
I just pass null for now (it's an optional field), I'll pass the tags in
https://github.com/DataDog/dd-trace-dotnet/pull/7715 once this is
merged, to make it easier to detect issues if any.

## Reason for change

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

https://github.com/DataDog/libdatadog/releases/tag/v25.0.0

<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->