Add PR-level performance quality gates (#7967)
## Summary of changes

- Adds PR-level gates to dd-trace-dotnet with:
- A regression threshold of 20%:
https://github.com/DataDog/benchmarking-platform/pull/224/files#diff-e2c6df715b4e55dca3365e30ac3801dd5d9ca5c212e18ed45ead5e2cff811077R6
- A `check-big-regressions-no-ignored-benchmarks` CI job that won't
block merges (since `allow-failure` is set to true).
- Update `.gitlab-ci.yml`'s `microbenchmarks`:
- Set `strategy: depends` to have `microbenchmarks` fail if
`check-big-regressions` (the PR gate) fails.
- Set `allow_failure: false` to have `microbenchmarks` failures prevent
PRs from being merged.

## Reason for change

https://datadoghq.atlassian.net/browse/APMSP-2292

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

`run-benchmarks` has a known issue @NachoEchevarria recently mentioned
of jobs failing when AWS allocates a machine with a CPU different than
the one we use for benchmarking.

This happens rarely, but it might lead to some frustration when
attempting merges, because it would require manual retries.

An automatic retry strategy would require changes to `bp-infra` I'm
planning to do. But I think that we can get this PR merged as is,
especially considering that there are other jobs on dd-trace-dotnet's CI
that require manual retries sometimes.

If failures prove to be too disruptive, we can always set
`allow_failure` to `true` and disable PR gates momentarily.


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->