Stop recreating `SpanEventsManager` (#7978)
## Summary of changes

Use the provided `SpanEventsManager` in `TracerManager` constructor

## Reason for change

While working on something else, @daniel-romano-DD noticed that we were
recreating the `SpanEventsManager` in the `TracerManager` constructor,
instead of using the value provided (which will have _just_ been created
previously. Can't see a good reason for it, it looks like a bug.

## Implementation details

Just use the provided value

## Test coverage

Probably benign so not covered by anything else, but meh