Stop storing tracer setting telemetry forever (#7914)
## Summary of changes

Stop storing the `TracerSettings` telemetry forever

## Reason for change

Back in v2, when we first implemented config telemetry, we had to store
the telemetry on `TracerSettings`, because a customer could
"reconfigure" the tracer with an _existing_ `TracerSettings` object,
thereby overriding the "latest" config. Put another way, the config
recorded when the `TracerSettings` was created, wasn't _necessarily_ the
config actually used.

As of v3, and particularly with the `MutableSettings` split, this is no
longer the case. The `TracerSettings` object is created only once, and
the results are immutable, so there's no need to keep the config around
"just in case".

## Implementation details

Instead of creating a dedicated `ConfigurationTelemetry` object
associated with the `TracerSettings`, and then copying it into the
collector, just write directly to the collector.

## Test coverage

Should be covered by existing, but updated some tests (and made one of
the constructor overloads easier to use in tests)

## Other details

Should have a (tiny) impact on startup time and overall memory use