Ensure we run the BenchmarksOpenTelemetryApi benchmarks on master (#8012)
## Summary of changes

Make sure we run the "OTel API only" benchmarks on master

## Reason for change

We created "OTel API only" benchmarks in #6968 to compare with our "DD
instrumented OTel API" benchmarks, but we're currently never running
them, so we have no data. This ensures we run them on master (only) so
we have comparison data.

## Implementation details

Check if we're on a PR (which we are already doing). If not, then run
the OTel API only benchmarks.

## Test coverage

By its nature, hard to confirm it's working exactly as expected. Will
confirm after merge